seperated things into hypervisor/gl.go
